refacto(libccc/string): Moved stringascii.h and stringutf8.h to text/ dir + renamed all include of string.h in stringascii dir to stringascii.h
