+++ hdr/libccc/text/stringutf8.h
