+++ src/stringascii/duplicate.c
@@ -1,7 +1,7 @@
 
 #include "libccc/char.h"
 #include "libccc/memory.h"
-#include "libccc/string.h"
+#include "libccc/text/stringascii.h"
 
 #if LIBCONFIG_USE_STD_FUNCTIONS_ALWAYS
 #include <string.h>
