+++ src/stringascii/join.c
@@ -1,5 +1,5 @@
 
-#include "libccc/string.h"
+#include "libccc/text/stringascii.h"
 #include "libccc/stringarray.h"
 
 #include LIBCONFIG_ERROR_INCLUDE
