+++ src/stringascii/length_n.c
@@ -1,5 +1,5 @@
 
-#include "libccc/string.h"
+#include "libccc/text/stringascii.h"
 
 #if LIBCONFIG_USE_STD_FUNCTIONS_ALWAYS
 #include <string.h>
